1. Missed an int when parsing damage which stopped multi mob attacks from working. 2. Set command for stats 3. Moved exp and level up handling to CharacterStats

+++ Common/Game/Entity/Meta/Extensions.cs
@@ -50,7 +50,7 @@ namespace NineToFive.Game.Entity.Meta {
                 case UserAbility.MP:         return user.CharacterStat.MP;
                 case UserAbility.MaxMP:      return user.CharacterStat.MaxMP;
                 case UserAbility.AP:         return user.CharacterStat.AP;
-                case UserAbility.Exp:        return user.CharacterStat.Exp;
+                case UserAbility.Exp:        return (int) user.CharacterStat.Exp;
                 case UserAbility.Popularity: return user.CharacterStat.Popularity;
                 case UserAbility.FieldId:    return user.CharacterStat.FieldId;
                 case UserAbility.Money:      return (int) user.Money;
